Add files via upload
These are the default items loaded for the Wheel Of Luck.  The WheelOfLuck-Default.json is a blank list of default items and their values.  The WheelOfLuck-Included.json has an example list of 106 items ready to give away.  Each item has values from the Class RustItem.  To include an item in the list set the value to true in the .json file
ItemIncluded = true
IncludeCategory = true
The multiplier setting is to set how many of the item the player will get (Default stack value Stackable) so if you put say 1 for a weapon they would get 1 x the stackable value in this case is 1 they would get 1 weapon. If you did something like wood the default stack is 1000 so if you set the ItemMultiplier to 3 they would get 3 x 1000 = 3000. I also am included an application that can generate the item list via a win app gui this will soon be uploaded.  If you have any questions send me a message!

public class RustItem
{
    public int ItemId { get; set; }
    public bool IncludeItem { get; set; }
    public string ItemImage { get; set; }
    public int ItemMultiplier { get; set; }
    public int AmountOfItemToGive { get; set; }
    public string ShortName { get; set; }
    public string Name { get; set; }
    public string Description { get; set; }
    public int Stackable { get; set; }
    public string Category { get; set; }
    public bool IncludeCategory { get; set; }
}